Use endpoint instead of model in the inference API
