+++ specification/_json_spec/inference.delete.json
@@ -1,8 +1,8 @@
 {
-  "inference.get_model": {
+  "inference.delete": {
     "documentation": {
-      "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/get-inference-api.html",
-      "description": "Get a model in the Inference API"
+      "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/delete-inference-api.html",
+      "description": "Delete an inference endpoint"
     },
     "stability": "experimental",
     "visibility": "public",
@@ -12,12 +12,8 @@
     "url": {
       "paths": [
         {
-          "path": "/_inference",
-          "methods": ["GET"]
-        },
-        {
           "path": "/_inference/{inference_id}",
-          "methods": ["GET"],
+          "methods": ["DELETE"],
           "parts": {
             "inference_id": {
               "type": "string",
@@ -27,7 +23,7 @@
         },
         {
           "path": "/_inference/{task_type}/{inference_id}",
-          "methods": ["GET"],
+          "methods": ["DELETE"],
           "parts": {
             "task_type": {
               "type": "string",
