+++ specification/_json_spec/inference.inference.json
@@ -2,7 +2,7 @@
   "inference.inference": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/post-inference-api.html",
-      "description": "Perform inference on a model"
+      "description": "Perform inference"
     },
     "stability": "experimental",
     "visibility": "public",
