+++ specification/_json_spec/inference.put.json
@@ -1,8 +1,8 @@
 {
-  "inference.put_model": {
+  "inference.put": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/put-inference-api.html",
-      "description": "Configure a model for use in the Inference API"
+      "description": "Configure an inference endpoint for use in the Inference API"
     },
     "stability": "experimental",
     "visibility": "public",
@@ -39,7 +39,7 @@
       ]
     },
     "body": {
-      "description": "The model's task and service settings"
+      "description": "The endpoint's task and service settings"
     }
   }
 }
