+++ specification/inference/delete/DeleteResponse.ts
