+++ specification/inference/put/PutResponse.ts
@@ -17,8 +17,8 @@
  * under the License.
  */
 
-import { ModelConfigContainer } from '@inference/_types/Services'
+import { InferenceEndpointContainer } from '@inference/_types/Services'
 
 export class Response {
-  body: ModelConfigContainer
+  body: InferenceEndpointContainer
 }
